[#11] 음식 기록 수정 삭제 api

+++ src/main/java/com/example/demo/domain/DB/DietList.java
@@ -29,7 +29,7 @@ public class DietList {
 
 
     @Column(name = "food_weight")
-    private Long foodWeight;
+    private Double foodWeight;
 
     @Column(name="food_kcal")
     private Long foodKcal;
